backend/pyproject.toml
[project]
name = "monitor-app"
version = "0.1.0"
description = "カメラ撮影した生体情報モニターの画像からバイタル情報を抽出するWebアプリ"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "anthropic>=0.40.0",
    "pydantic-settings>=2.6.0",
    "Pillow>=11.0.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "psycopg2-binary>=2.9.11",
]

[tool.uv]
package = false
